requirements.txt
numpy~=2.3.5
torch~=2.9.1
scikit-learn~=1.8.0
pillow~=12.0.0
matplotlib~=3.10.8
tqdm~=4.67.1
torchvision~=0.24.1
pandas~=2.3.3